requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-core
azure-storage-blob
azure-identity
azure-servicebus
openai>=1.0.0
requests
